[Feat] #44 - PacingModel 에서 빼와서 ListViewModel 생성
